// src/app/modules/public/model/video.ts
export class videoData {
    VideoID: any;
    Title: any;
    VideoType: any;
    Description: any;
    Embed_Code: any;
    ExpiresDate: any;
    Levels: any;
    Orders: any;
    Prerecord_Video_Image: any;
    ResourceUrl: any;
    SortOrder: any;
    IsActive: any;
    StartTime: any;
    StartDate: any;
    EndDate: any;
    EndTime: any;
    CreatedBy: any;
    CreatedDate: any;
    ModifiedBy: any;
    ModifiedDate: any;
    videoInfo: any;
 }
 
 export class playVideo {
    Prerecord_Video_Image: any;
    Embed_Code: any;
 }